scripts/requirements.txt
python-pptx>=0.6.23
pypdf>=5.4.0
